requirements_web.txt
# requirements_web.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
